pyproject.toml
[project]
name = "chimerapy-rerun-node"
version = "0.0.3"
description = "Log ChimeraPy data to a rerun server"
authors = [
    {name = "Umesh Timalsina", email="umesh.timalsina@vanderbilt.edu"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">3.6"

keywords = ["multimodal", "data", "distributed", "streaming", "real-time"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    'chimerapy-engine',
    'chimerapy-orchestrator',
    'rerun-sdk',
]

[project.optional-dependencies]
test = [
    'pytest',
    'pytest-anyio',
    'coveralls',
    'pre-commit',
    'coverage[toml]',
    'bump2version'
]

examples = [
    "transformers",
    "timm"
]

[project.urls]
homepath = "https://github.com/chimerapy/Orchestrator"
documentation = "https://github.com/chimerapy/Orchestrator"
repository = "https://github.com/chimerapy/Orchestrator"

# Entrypoint
[project.scripts]
cp-orchestrator = "chimerapy.orchestrator.cli.__main__:run"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# https://setuptools.pypa.io/en/stable/userguide/datafiles.html
[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
chimerapy = ['orchestrator/**/*.json']

[project.entry-points."chimerapy.orchestrator.nodes_registry"]
get_nodes_registry = "chimerapy.rerun_node:register_nodes_metadata"


[tool.setuptools.packages.find]
where = ["."]

[tool.coverage]
[tool.coverage.run]
omit = ["*/tests/*", "*/__init__.py", "*/__main__.py", "*/__version__.py"]

[tool.ruff]
ignore = ["E501"]
select = ["E", "W", "F", "C", "B", "I"]
ignore-init-module-imports = true
fixable = ["I001"]  # isort fix only

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
